+++ xlwings_utils/__init__.py
@@ -1,2 +1,5 @@
 from .xlwings_utils import *
 from .xlwings_utils import __version__
+from .dropbox import *
+from .nextcloud import *
+from .local import *